create context for userid in cart service

+++ src/main/java/com/shopsphere/cart_service/dto/CartDTO.java
@@ -4,6 +4,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -18,5 +19,5 @@ public class CartDTO {
     private List<CartItemDTO> cartItems;
 
     @Schema(description = "total cart price in USD", example = "150.00")
-    private Double totalCartPrice;
+    private BigDecimal totalCartPrice;
 }
